+++ androidApp/src/main/java/com/rld/justlisten/android/ui/bottombars/bottombarnav/Level1BottomBar.kt
@@ -1,4 +1,4 @@
-package com.rld.justlisten.android.ui.bottombars
+package com.rld.justlisten.android.ui.bottombars.bottombarnav
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.BoxWithConstraints
