+++ androidApp/src/main/java/com/rld/justlisten/android/ui/bottombars/playbar/components/more/PlayBarMoreAction.kt
@@ -17,7 +17,6 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.rld.justlisten.android.R
-import com.rld.justlisten.android.ui.bottombars.sheetcontent.BottomSheetScreen
 
 @Composable
 fun PlayBarMoreAction(
