+++ androidApp/src/main/java/com/rld/justlisten/android/ui/bottombars/sheets/BottomSheetScreen.kt
@@ -1,4 +1,4 @@
-package com.rld.justlisten.android.ui.bottombars.sheetcontent
+package com.rld.justlisten.android.ui.bottombars.sheets
 
 sealed class BottomSheetScreen {
     object More: BottomSheetScreen()
